+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Fix empty cookie attributes being set to `Key=` instead of `Key`
   ([#5084](https://github.com/mitmproxy/mitmproxy/pull/5084), @Speedlulu)
+* Scripts with relative paths are now loaded relative to the config file and not where the command is ran
 
 
 ## 14 November 2023: mitmproxy 10.1.5